+++ src/Factory.php
@@ -9,6 +9,8 @@ use Phlib\Mail\Exception\RuntimeException;
 
 class Factory
 {
+    private const RECEIVED_REGEX = '/^(.*);\s*((?:\w{3},\s)?\d{1,2}\s\w{3}\s\d{4}\s\d{2}:\d{2}(?::\d{2})?\s(?:\+|-)\d{4})/';
+
     /**
      * @var bool
      */
@@ -151,6 +153,15 @@ class Factory
 
                 try {
                     switch (strtolower($headerKey)) {
+                        case 'date':
+                            $date = new \DateTimeImmutable($headerText);
+                            $mail->setOriginationDate($date);
+                            break;
+                        case 'received':
+                            if (preg_match(self::RECEIVED_REGEX, $headerText, $received) > 0) {
+                                $mail->addReceived(trim($received[1]), new \DateTimeImmutable($received[2]));
+                            }
+                            break;
                         case 'return-path':
                         case 'from':
                         case 'reply-to':
@@ -179,6 +190,14 @@ class Factory
                         case 'message-id':
                             $messageId = $this->parseEmailAddresses($headerText);
                             $mail->setMessageId($messageId[0]['address']);
+                            break;
+                        case 'in-reply-to':
+                            $addresses = $this->parseEmailAddresses($headerText);
+                            $mail->setInReplyTo(array_column($addresses, 'address'));
+                            break;
+                        case 'references':
+                            $addresses = $this->parseEmailAddresses($headerText);
+                            $mail->setReferences(array_column($addresses, 'address'));
                             break;
                         case 'subject':
                             $mail->setSubject($headerText);